Did some bugfixes in lobby.js.
- In disconnect: we previously used lobby_members.indexOf(socket.id), but lobby_members is a list of usernames, not socketIDs, so I changed this. I also added a line to delete the entry for that socket in socket_usernames.
- In username: added a check for that socketID/username already existing.
- In join room: added a check for the room already existing, because previously multiple users joining the same room would result in that room being listed several times. Also removed that user from lobby

I'm having some trouble with disconnects when I have multiple users - the user that disconnects is shown as the only one that remains in the lobby, when they should be the one that is removed. May have to do with the fact that all the connections are from me, and this would work if the connections were from different computers.